requirements.txt
openai
python-dotenv
qdrant-client
streamlit
streamlit-audiorecorder
pydub
pycaret
pathlib
os